refactor(*): extract most of the main routine from the constructor into a new async instance method, `_main`
