+++ src/instanceMethods.js
@@ -2,3 +2,4 @@ export * from './instanceMethods/hideLoading'
 export * from './instanceMethods/getInput'
 export * from './instanceMethods/enable-disable-elements'
 export * from './instanceMethods/show-reset-validation-error'
+export * from './instanceMethods/_main'
